week3/pytorch-titanic/requirements.txt
torch
pandas
scikit-learn
argparse
matplotlib
tqdm